+++ stories/components/helpers/createApiStory.js
@@ -1,0 +1,8 @@
+import React from 'react'
+
+export default (story, Component) => {
+  story.add('API', () => <div />, {
+    info: { inline: true, source: false, propTables: [Component] },
+    options: { showPanel: false },
+  })
+}