+++ stories/helpers/createApiStory.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import { storiesOf } from '@storybook/react'
-
-export default (title = '', Component) => {
-  storiesOf(title, module).add('API', () => <div />, {
-    info: { inline: true, source: false, propTables: [Component] },
-    options: { showPanel: false },
-  })
-}